requirements.txt
Flask==1.0.2
requests==2.18.4
beautifulsoup4==4.6.0
django-htmlmin==0.10.0
gunicorn==19.8.1
flask-compress==1.4.0
